refactor(mcp): hoist duplicated omni_video_generation plumbing into mcp-common
The MCP argument parsing/validation and response rendering for the
omni_video_generation tool were duplicated verbatim between
mcp-omni-go/handlers.go and mcp-gemini-go/omni_handlers.go (~360 lines).
This is the class of duplication that forced the Phase-2a input-hardening
bot fixes (integer temp/top_p, image-count-before-read, 20 MiB inline cap,
IsDir reject, MIME-derived extension) to be applied twice.

Hoist the shared logic into mcp-common/omni_args.go behind two functions,
following the existing mcp-common args.go convention of taking the raw
map[string]interface{} and returning typed values / plain errors (no MCP
framework types leak into the shared library):

  - ParseOmniToolArgs(args, cfg) -> OmniToolArgs {Params, OutputDir,
    GCSBucketURI} performs all client-side validation.
  - RenderOmniResult(ctx, result, outputDir, gcsBucketURI) persists each
    video and builds the user-facing text response.

Both handlers now call ParseOmniToolArgs -> GenerateOmniVideo ->
RenderOmniResult and shrink to thin wrappers (~80 lines) whose only
remaining server-specific concerns are the OTEL span (needs per-server
serviceName) and the mcp.CallTool signature. The tool registration block
stays in each main.go to keep mcp-common free of the mcp-go framework
dependency (declarative config, never carried a behavior-affecting fix).

Behavior-preserving: parameter names, validation error strings, output/NOTE
text, filenames, and the OMNI_SIGNED_URL_EXPIRY_HOURS env var are byte-for-
byte unchanged (the moved helpers are byte-identical to the originals). The
helper unit tests move from mcp-omni-go/handlers_test.go into
mcp-common/omni_args_test.go and are joined by direct coverage of
ParseOmniToolArgs and RenderOmniResult.
