+++ experiments/mcp-genmedia/mcp-genmedia-go/mcp-common/omni_args_test.go
@@ -1,0 +1,441 @@
+// Copyright 2025 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package common
+
+import (
+	"context"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+)
+
+// TestParseMediaRefsGCSURI verifies gs:// entries are passed through by URI (no
+// file read) with a MIME type inferred from the extension.
+func TestParseMediaRefsGCSURI(t *testing.T) {
+	refs, err := parseMediaRefs([]interface{}{"gs://bucket/cat.png"}, "image")
+	if err != nil {
+		t.Fatalf("parseMediaRefs returned error: %v", err)
+	}
+	if len(refs) != 1 {
+		t.Fatalf("len(refs) = %d, want 1", len(refs))
+	}
+	if refs[0].URI != "gs://bucket/cat.png" {
+		t.Errorf("URI = %q, want gs://bucket/cat.png", refs[0].URI)
+	}
+	if len(refs[0].Data) != 0 {
+		t.Errorf("gs:// ref should carry no inline data, got %d bytes", len(refs[0].Data))
+	}
+	if refs[0].MimeType != "image/png" {
+		t.Errorf("MimeType = %q, want image/png", refs[0].MimeType)
+	}
+}
+
+// TestParseMediaRefsLocalFile verifies a local path is read into inline bytes.
+func TestParseMediaRefsLocalFile(t *testing.T) {
+	dir := t.TempDir()
+	path := filepath.Join(dir, "clip.mp4")
+	want := []byte("fake-mp4-bytes")
+	if err := os.WriteFile(path, want, 0o600); err != nil {
+		t.Fatalf("failed to write temp file: %v", err)
+	}
+
+	refs, err := parseMediaRefs([]interface{}{path}, "video")
+	if err != nil {
+		t.Fatalf("parseMediaRefs returned error: %v", err)
+	}
+	if len(refs) != 1 {
+		t.Fatalf("len(refs) = %d, want 1", len(refs))
+	}
+	if string(refs[0].Data) != string(want) {
+		t.Errorf("Data = %q, want %q", refs[0].Data, want)
+	}
+	if refs[0].URI != "" {
+		t.Errorf("local ref should carry no URI, got %q", refs[0].URI)
+	}
+	if refs[0].MimeType != "video/mp4" {
+		t.Errorf("MimeType = %q, want video/mp4", refs[0].MimeType)
+	}
+}
+
+func TestParseMediaRefsErrors(t *testing.T) {
+	if _, err := parseMediaRefs("not-an-array", "image"); err == nil {
+		t.Error("expected error for a non-array argument")
+	}
+	if _, err := parseMediaRefs([]interface{}{""}, "image"); err == nil {
+		t.Error("expected error for an empty-string entry")
+	}
+	if _, err := parseMediaRefs([]interface{}{123}, "image"); err == nil {
+		t.Error("expected error for a non-string entry")
+	}
+	if _, err := parseMediaRefs([]interface{}{"/no/such/file.png"}, "image"); err == nil {
+		t.Error("expected error for an unreadable local file")
+	}
+}
+
+// TestParseMediaRefsDirectoryRejected guards BOT-2: a directory path is rejected
+// before any size check / read (info.IsDir()).
+func TestParseMediaRefsDirectoryRejected(t *testing.T) {
+	dir := t.TempDir()
+	_, err := parseMediaRefs([]interface{}{dir}, "image")
+	if err == nil {
+		t.Fatal("expected an error for a directory path")
+	}
+	if !strings.Contains(err.Error(), "is a directory") {
+		t.Errorf("error should mention it is a directory, got: %v", err)
+	}
+}
+
+// TestParseMediaRefsNilYieldsNil confirms an absent argument produces no refs.
+func TestParseMediaRefsNilYieldsNil(t *testing.T) {
+	refs, err := parseMediaRefs(nil, "image")
+	if err != nil {
+		t.Fatalf("parseMediaRefs(nil) returned error: %v", err)
+	}
+	if refs != nil {
+		t.Errorf("parseMediaRefs(nil) = %v, want nil", refs)
+	}
+}
+
+func TestInferMediaMimeType(t *testing.T) {
+	cases := map[string]string{
+		"a.png":       "image/png",
+		"a.JPG":       "image/jpeg",
+		"a.jpeg":      "image/jpeg",
+		"a.webp":      "image/webp",
+		"clip.mp4":    "video/mp4",
+		"clip.mov":    "video/quicktime",
+		"clip.webm":   "video/webm",
+		"unknown.bin": "application/octet-stream",
+	}
+	for name, want := range cases {
+		if got := inferMediaMimeType(name); got != want {
+			t.Errorf("inferMediaMimeType(%q) = %q, want %q", name, got, want)
+		}
+	}
+}
+
+// TestVideoExtForMimeType guards BOT-1: the saved filename's extension is derived
+// from the resolved MIME type (defaulting to .mp4), not hardcoded.
+func TestVideoExtForMimeType(t *testing.T) {
+	cases := map[string]string{
+		"video/mp4":       ".mp4",
+		"video/webm":      ".webm",
+		"video/quicktime": ".mov",
+		"video/mpeg":      ".mpeg",
+		"video/3gpp":      ".3gp",
+		"":                ".mp4",
+		"application/pdf": ".mp4",
+	}
+	for mime, want := range cases {
+		if got := videoExtForMimeType(mime); got != want {
+			t.Errorf("videoExtForMimeType(%q) = %q, want %q", mime, got, want)
+		}
+	}
+}
+
+func TestToInt(t *testing.T) {
+	if n, err := toInt(float64(3)); err != nil || n != 3 {
+		t.Errorf("toInt(3.0) = %d, %v; want 3, nil", n, err)
+	}
+	if n, err := toInt(2); err != nil || n != 2 {
+		t.Errorf("toInt(2) = %d, %v; want 2, nil", n, err)
+	}
+	if _, err := toInt(float64(2.5)); err == nil {
+		t.Error("expected error for a non-integer float")
+	}
+	if _, err := toInt("nope"); err == nil {
+		t.Error("expected error for a string argument")
+	}
+}
+
+func TestParseOptionalFloatInRange(t *testing.T) {
+	args := map[string]interface{}{
+		"temperature": float64(1.5),
+		"too_high":    float64(9),
+		"not_number":  "x",
+	}
+
+	// Absent key -> nil, no error.
+	v, err := parseOptionalFloatInRange(args, "missing", 0, 2)
+	if err != nil || v != nil {
+		t.Errorf("absent key: got %v, %v; want nil, nil", v, err)
+	}
+
+	// In-range value.
+	v, err = parseOptionalFloatInRange(args, "temperature", 0, 2)
+	if err != nil || v == nil || *v != 1.5 {
+		t.Errorf("in-range: got %v, %v; want 1.5, nil", v, err)
+	}
+
+	// Out-of-range value.
+	if _, err := parseOptionalFloatInRange(args, "too_high", 0, 2); err == nil {
+		t.Error("expected error for an out-of-range value")
+	}
+
+	// Wrong type.
+	if _, err := parseOptionalFloatInRange(args, "not_number", 0, 2); err == nil {
+		t.Error("expected error for a non-numeric value")
+	}
+}
+
+// TestParseOptionalFloatInRangeIntegers guards BOT-1: integer JSON numbers (which
+// may surface as float64, int, or int64) must be accepted for temperature/top_p,
+// not rejected as "must be a number".
+func TestParseOptionalFloatInRangeIntegers(t *testing.T) {
+	cases := map[string]interface{}{
+		"float64_int": float64(1), // JSON "1" decoded to float64
+		"int":         1,
+		"int64":       int64(1),
+	}
+	for name, raw := range cases {
+		t.Run(name, func(t *testing.T) {
+			// temperature range [0,2]
+			v, err := parseOptionalFloatInRange(map[string]interface{}{"temperature": raw}, "temperature", 0, 2)
+			if err != nil {
+				t.Fatalf("integer temperature %v (%T) rejected: %v", raw, raw, err)
+			}
+			if v == nil || *v != 1 {
+				t.Fatalf("temperature = %v, want 1", v)
+			}
+			// top_p range [0,1]
+			p, err := parseOptionalFloatInRange(map[string]interface{}{"top_p": raw}, "top_p", 0, 1)
+			if err != nil {
+				t.Fatalf("integer top_p %v (%T) rejected: %v", raw, raw, err)
+			}
+			if p == nil || *p != 1 {
+				t.Fatalf("top_p = %v, want 1", p)
+			}
+		})
+	}
+}
+
+// TestParseMediaRefsOversized guards BOT-3: a local file above the inline cap is
+// rejected with a hint to use a gs:// URI, and is not read fully into memory.
+func TestParseMediaRefsOversized(t *testing.T) {
+	dir := t.TempDir()
+	path := filepath.Join(dir, "big.mp4")
+	f, err := os.Create(path)
+	if err != nil {
+		t.Fatalf("create: %v", err)
+	}
+	// Sparse file just over the cap (no large allocation).
+	if err := f.Truncate(maxInlineMediaBytes + 1); err != nil {
+		t.Fatalf("truncate: %v", err)
+	}
+	f.Close()
+
+	_, err = parseMediaRefs([]interface{}{path}, "video")
+	if err == nil {
+		t.Fatal("expected an error for an oversized local file")
+	}
+	if !strings.Contains(err.Error(), "gs://") {
+		t.Errorf("error should suggest a gs:// URI, got: %v", err)
+	}
+}
+
+// TestParseOmniToolArgsHappyPath exercises the full argument surface both servers
+// share: prompt trimming, model default, media refs, sample_count, and the
+// GENMEDIA_BUCKET fallback for gcs_bucket_uri.
+func TestParseOmniToolArgsHappyPath(t *testing.T) {
+	cfg := &Config{GenmediaBucket: "my-bucket"}
+	args := map[string]interface{}{
+		"prompt":           "  a cat surfing  ",
+		"images":           []interface{}{"gs://bucket/cat.png"},
+		"sample_count":     float64(2),
+		"temperature":      float64(1),
+		"top_p":            float64(0.5),
+		"output_directory": "/tmp/out",
+	}
+
+	parsed, err := ParseOmniToolArgs(args, cfg)
+	if err != nil {
+		t.Fatalf("ParseOmniToolArgs returned error: %v", err)
+	}
+	if parsed.Params.Prompt != "a cat surfing" {
+		t.Errorf("Prompt = %q, want trimmed %q", parsed.Params.Prompt, "a cat surfing")
+	}
+	if parsed.Params.Model != DefaultOmniModel {
+		t.Errorf("Model = %q, want default %q", parsed.Params.Model, DefaultOmniModel)
+	}
+	if len(parsed.Params.Images) != 1 || parsed.Params.Images[0].URI != "gs://bucket/cat.png" {
+		t.Errorf("Images = %+v, want one gs:// image ref", parsed.Params.Images)
+	}
+	if parsed.Params.SampleCount != 2 {
+		t.Errorf("SampleCount = %d, want 2", parsed.Params.SampleCount)
+	}
+	if parsed.Params.Temperature == nil || *parsed.Params.Temperature != 1 {
+		t.Errorf("Temperature = %v, want 1", parsed.Params.Temperature)
+	}
+	if parsed.Params.TopP == nil || *parsed.Params.TopP != 0.5 {
+		t.Errorf("TopP = %v, want 0.5", parsed.Params.TopP)
+	}
+	if parsed.OutputDir != "/tmp/out" {
+		t.Errorf("OutputDir = %q, want /tmp/out", parsed.OutputDir)
+	}
+	if parsed.GCSBucketURI != "my-bucket/omni_outputs/" {
+		t.Errorf("GCSBucketURI = %q, want GENMEDIA_BUCKET fallback", parsed.GCSBucketURI)
+	}
+}
+
+// TestParseOmniToolArgsExplicitBucketWins confirms an explicit gcs_bucket_uri
+// takes precedence over the GENMEDIA_BUCKET fallback.
+func TestParseOmniToolArgsExplicitBucketWins(t *testing.T) {
+	cfg := &Config{GenmediaBucket: "fallback-bucket"}
+	parsed, err := ParseOmniToolArgs(map[string]interface{}{
+		"prompt":         "hi",
+		"gcs_bucket_uri": "explicit-bucket/dir/",
+	}, cfg)
+	if err != nil {
+		t.Fatalf("ParseOmniToolArgs returned error: %v", err)
+	}
+	if parsed.GCSBucketURI != "explicit-bucket/dir/" {
+		t.Errorf("GCSBucketURI = %q, want the explicit value", parsed.GCSBucketURI)
+	}
+}
+
+// TestParseOmniToolArgsDefaults verifies sample_count defaults to 1 and an absent
+// GENMEDIA_BUCKET leaves gcs_bucket_uri empty.
+func TestParseOmniToolArgsDefaults(t *testing.T) {
+	parsed, err := ParseOmniToolArgs(map[string]interface{}{"prompt": "hi"}, &Config{})
+	if err != nil {
+		t.Fatalf("ParseOmniToolArgs returned error: %v", err)
+	}
+	if parsed.Params.SampleCount != 1 {
+		t.Errorf("SampleCount = %d, want default 1", parsed.Params.SampleCount)
+	}
+	if parsed.GCSBucketURI != "" {
+		t.Errorf("GCSBucketURI = %q, want empty", parsed.GCSBucketURI)
+	}
+	if parsed.Params.Temperature != nil || parsed.Params.TopP != nil {
+		t.Errorf("unset temperature/top_p should be nil, got %v/%v", parsed.Params.Temperature, parsed.Params.TopP)
+	}
+}
+
+func TestParseOmniToolArgsValidationErrors(t *testing.T) {
+	cfg := &Config{}
+	cases := map[string]struct {
+		args    map[string]interface{}
+		wantSub string
+	}{
+		"missing prompt": {
+			args:    map[string]interface{}{},
+			wantSub: "prompt must be a non-empty string and is required",
+		},
+		"blank prompt": {
+			args:    map[string]interface{}{"prompt": "   "},
+			wantSub: "prompt must be a non-empty string and is required",
+		},
+		"unsupported model": {
+			args:    map[string]interface{}{"prompt": "hi", "model": "not-a-real-model"},
+			wantSub: "unsupported model",
+		},
+		"too many images": {
+			args: map[string]interface{}{
+				"prompt": "hi",
+				"images": make([]interface{}, maxOmniImages+1),
+			},
+			wantSub: "too many images",
+		},
+		"non-integer sample_count": {
+			args:    map[string]interface{}{"prompt": "hi", "sample_count": float64(2.5)},
+			wantSub: "sample_count must be an integer",
+		},
+		"temperature out of range": {
+			args:    map[string]interface{}{"prompt": "hi", "temperature": float64(3)},
+			wantSub: "temperature must be in the range",
+		},
+	}
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			_, err := ParseOmniToolArgs(tc.args, cfg)
+			if err == nil {
+				t.Fatalf("expected an error containing %q, got nil", tc.wantSub)
+			}
+			if !strings.Contains(err.Error(), tc.wantSub) {
+				t.Errorf("error = %q, want it to contain %q", err.Error(), tc.wantSub)
+			}
+		})
+	}
+}
+
+// TestParseOmniToolArgsImageCountGuardBeforeRead confirms the image-count guard
+// fires before any file is read (BOT-2 ordering): a too-long list of nonexistent
+// paths still fails with the count error, not a file-read error.
+func TestParseOmniToolArgsImageCountGuardBeforeRead(t *testing.T) {
+	imgs := make([]interface{}, maxOmniImages+1)
+	for i := range imgs {
+		imgs[i] = "/no/such/file.png"
+	}
+	_, err := ParseOmniToolArgs(map[string]interface{}{"prompt": "hi", "images": imgs}, &Config{})
+	if err == nil || !strings.Contains(err.Error(), "too many images") {
+		t.Fatalf("want a too-many-images error before any read, got: %v", err)
+	}
+}
+
+// TestRenderOmniResultLocal exercises the shared response renderer end to end for
+// a local-only save (no GCS), asserting the filename extension, saved-path
+// summary, model text, and thought-step NOTE all appear.
+func TestRenderOmniResultLocal(t *testing.T) {
+	dir := t.TempDir()
+	result := &OmniResult{
+		Videos:         [][]byte{[]byte("mp4-bytes")},
+		VideoMimeTypes: []string{"video/mp4"},
+		Text:           "Here is your video.",
+		ThoughtSteps:   1,
+	}
+
+	msg, err := RenderOmniResult(context.Background(), result, dir, "")
+	if err != nil {
+		t.Fatalf("RenderOmniResult returned error: %v", err)
+	}
+	if !strings.Contains(msg, "Here is your video.") {
+		t.Errorf("message missing model text: %q", msg)
+	}
+	if !strings.Contains(msg, "reasoning (thought) step(s).") {
+		t.Errorf("message missing thought-step NOTE: %q", msg)
+	}
+	if !strings.Contains(msg, "Generated and saved 1 video(s):") {
+		t.Errorf("message missing saved summary: %q", msg)
+	}
+
+	entries, err := os.ReadDir(dir)
+	if err != nil {
+		t.Fatalf("ReadDir: %v", err)
+	}
+	if len(entries) != 1 {
+		t.Fatalf("expected 1 saved file, got %d", len(entries))
+	}
+	name := entries[0].Name()
+	if !strings.HasPrefix(name, "omni_") || !strings.HasSuffix(name, ".mp4") {
+		t.Errorf("saved filename = %q, want omni_*.mp4", name)
+	}
+}
+
+// TestRenderOmniResultNoDestination confirms the "none saved" summary is used
+// when neither an output directory nor a GCS bucket is provided.
+func TestRenderOmniResultNoDestination(t *testing.T) {
+	result := &OmniResult{
+		Videos:         [][]byte{[]byte("mp4-bytes")},
+		VideoMimeTypes: []string{"video/mp4"},
+	}
+	msg, err := RenderOmniResult(context.Background(), result, "", "")
+	if err != nil {
+		t.Fatalf("RenderOmniResult returned error: %v", err)
+	}
+	if !strings.Contains(msg, "none were saved (set output_directory or gcs_bucket_uri).") {
+		t.Errorf("message missing none-saved summary: %q", msg)
+	}
+}
